# provider-apps/midtrans/openapi.yaml
openapi: 3.0.3
info:
  title: Midtrans Emisell Payment Connector
  version: emisell-midtrans-v2.0.2
  description: Normalized Payment Connector contract used by the shared Midtrans runtime.
servers:
  - url: https://connector-runtime.emisell.com/partner/v1
    description: Runtime Dispatcher route; the submission ZIP itself is never executed.
tags:
  - name: Platform
  - name: Installation
  - name: Payment
  - name: Webhook
paths:
  /health:
    get:
      operationId: getHealth
      tags: [Platform]
      responses:
        "200": {description: Connector is ready}
  /capabilities:
    get:
      operationId: getCapabilities
      tags: [Platform]
      responses:
        "200": {description: Connector manifest and capabilities}
  /payment-methods/validate:
    post:
      operationId: validatePaymentMethod
      tags: [Payment]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedValidation"}
      responses:
        "200": {description: Payment method mapping is valid}
        "422": {$ref: "#/components/responses/ErrorResponse"}
  /hosted-payment-methods/validate:
    post:
      operationId: validateHostedPaymentMethods
      tags: [Payment]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedValidation"}
      responses:
        "200": {description: Hosted payment method allowlist can be enforced exactly}
        "422": {$ref: "#/components/responses/ErrorResponse"}
  /payments/validate:
    post:
      operationId: validatePayment
      tags: [Payment]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedValidation"}
      responses:
        "200": {description: Payment request is supported}
        "422": {$ref: "#/components/responses/ErrorResponse"}
  /installations/verify:
    post:
      operationId: verifyInstallation
      tags: [Installation]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InstallationInput"}
      responses:
        "200":
          description: Server Key, detected environment, and provider access are verified
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: "#/components/schemas/InstallationResult"}
        "401": {$ref: "#/components/responses/ErrorResponse"}
        "422": {$ref: "#/components/responses/ErrorResponse"}
  /installations/disable:
    post:
      operationId: disableInstallation
      tags: [Installation]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InstallationInput"}
      responses:
        "200": {description: Provider-side installation resources are disabled}
        "422": {$ref: "#/components/responses/ErrorResponse"}
  /payments/create:
    post:
      operationId: createPayment
      tags: [Payment]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PaymentInput"}
      responses:
        "200": {description: Normalized payment was created}
        "422": {$ref: "#/components/responses/ErrorResponse"}
        "502": {$ref: "#/components/responses/ErrorResponse"}
  /payments/get:
    post:
      operationId: getPayment
      tags: [Payment]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PaymentLookup"}
      responses:
        "200": {description: Current normalized payment state}
        "404": {$ref: "#/components/responses/ErrorResponse"}
        "502": {$ref: "#/components/responses/ErrorResponse"}
  /webhooks/normalize:
    post:
      operationId: normalizeWebhook
      tags: [Webhook]
      description: Verifies the Midtrans signature key and returns one normalized event.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WebhookInput"}
      responses:
        "200": {description: Normalized webhook event}
        "401": {$ref: "#/components/responses/ErrorResponse"}
        "422": {$ref: "#/components/responses/ErrorResponse"}
components:
  securitySchemes:
    runtimeToken:
      type: http
      scheme: bearer
  schemas:
    VersionedValidation:
      type: object
      required: [provider_code, provider_version, input]
      properties:
        provider_code: {type: string, example: midtrans}
        provider_version: {type: string, example: emisell-midtrans-v2.0.2}
        input: {type: object, additionalProperties: true}
    InstallationInput:
      type: object
      required: [provider_code, provider_version, environment, credentials]
      properties:
        provider_code: {type: string, example: midtrans}
        provider_version: {type: string, example: emisell-midtrans-v2.0.2}
        environment: {type: string, enum: [live, sandbox]}
        credentials: {type: object, additionalProperties: {type: string}, writeOnly: true}
        public_webhook_url: {type: string, format: uri}
    InstallationResult:
      type: object
      required: [connector_id, environment, webhook_ready]
      properties:
        connector_id: {type: string, example: "midtrans:ins_01k3..."}
        environment: {type: string, enum: [sandbox, live]}
        webhook_ready: {type: boolean}
        payment_method_availability:
          type: array
          description: Optional short-lived runtime evidence; omitted when Midtrans does not expose per-channel health.
          items: {$ref: "#/components/schemas/PaymentMethodAvailability"}
    PaymentMethodAvailability:
      type: object
      required: [payment_method_code, status]
      properties:
        payment_method_code: {type: string, example: qris}
        status: {type: string, enum: [AVAILABLE, UNAVAILABLE]}
        reason: {type: string, example: PROVIDER_MAINTENANCE}
    PaymentInput:
      allOf:
        - {$ref: "#/components/schemas/InstallationInput"}
        - type: object
          required: [local_payment_id, amount, currency, idempotency_key]
          properties:
            local_payment_id: {type: string}
            checkout_mode: {type: string, enum: [direct, provider_hosted]}
            payment_method_code: {type: string}
            allowed_payment_methods:
              type: array
              description: Active installation mappings converted to the exact Midtrans Snap enabled_payments list.
              items:
                type: object
                required: [payment_method_code, provider_method, provider_method_type, provider_channel_code]
                properties:
                  payment_method_code: {type: string, example: qris}
                  provider_method: {type: string, example: real_time_payment}
                  provider_method_type: {type: string, example: qris}
                  provider_channel_code: {type: string, example: other_qris}
            amount: {type: integer, format: int64, minimum: 1, description: Whole IDR rupiah; 10000 means Rp10.000}
            currency: {type: string, example: IDR}
            idempotency_key: {type: string}
    PaymentLookup:
      allOf:
        - {$ref: "#/components/schemas/InstallationInput"}
        - type: object
          required: [payment_id]
          properties:
            payment_id: {type: string}
    WebhookInput:
      type: object
      required: [provider_code, provider_version, headers, body]
      properties:
        provider_code: {type: string, example: midtrans}
        provider_version: {type: string, example: emisell-midtrans-v2.0.2}
        headers: {type: object, additionalProperties: {type: string}}
        body: {type: string, format: byte}
  responses:
    ErrorResponse:
      description: Normalized connector error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string}
                  message: {type: string}
security:
  - runtimeToken: []